requirements.txt
attrs==21.2.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Cython==0.29.26
elasticsearch==7.13.1
Flask==1.1.2
gevent==20.12.0
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-bigquery==2.13.1
google-cloud-core==1.5.0
google-cloud-logging==2.0.2
google-cloud-pubsub==2.2.0
google-cloud-storage==1.36.2
google-crc32c==1.3.0
google-resumable-media==1.3.3
googleapis-common-protos==1.52.0
greenlet==0.4.17
grpc-google-iam-v1==0.12.3
grpcio==1.34.0
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
libcst==0.3.16
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pluggy==0.13.1
proto-plus==1.13.0
protobuf==3.14.0
psycopg2-binary==2.8.6
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pymongo==4.0.1
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
rsa==4.6
schedule==0.6.0
six==1.15.0
SQLAlchemy==1.3.22
supervisor==4.2.1
toml==0.10.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.2
vedis==0.7.1
Werkzeug==1.0.1
zope.event==4.5.0
zope.interface==5.2.0
sparkpost==1.3.10
dateparser==1.1.1
Cerberus==1.3.4
beautifulsoup4==4.12.2